% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{read_more}
\alias{read_more}
\title{Add "Read More" Separator}
\usage{
read_more()
}
\value{
Read More tag
}
\description{
This function allows you to insert a Read More tag
into your post.
}
\details{
used in your inline code wherever you want in your index.Rmd
}
\examples{
goodpress::read_more()
}
