% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wp_categories}
\alias{wp_categories}
\title{Get WordPress Categories}
\usage{
wp_categories(wordpress_url)
}
\arguments{
\item{wordpress_url}{URL to your website}
}
\value{
A data frame with category ids and names
}
\description{
This function helps you get your WordPress categories
details so that you can use for your YAML header.
}
\examples{
wordpress_url <- "https://rmd-wordpress.eu" # replace with your own (test) website
wp_categories(wordpress_url)
}
