% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{wp_post}
\alias{wp_post}
\title{Post a post to Wordpress}
\usage{
wp_post(post_folder, wordpress_url)
}
\arguments{
\item{post_folder}{Path to folder where the post (index.Rmd, index.md, images)
lives}

\item{wordpress_url}{URL to your website}
}
\value{
URL to the post (invisibly)
}
\description{
Post a post to Wordpress
}
\details{
To set the status of the post (One of: "publish", "future", "draft",
"pending", "private") use a YAML field e.g. \code{status: "draft"}.

The package cannot handle private posts with password, only private posts that
are visible to admins and editors only. You could create a private post, and
then from the WordPress interface make it visible with password. Make it private again
before trying to update the post with the R package.
}
\examples{
\dontrun{
# this requires authentication!
post_folder <- system.file(file.path("post-example"), package = "goodpress")
tmp_post_folder <- file.path(tempdir(), "post")
dir.create(tmp_post_folder)
file.copy(dir(post_folder, full.names = TRUE), tmp_post_folder)
wordpress_url <- "https://rmd-wordpress.eu" # replace with your own (test) website
wp_post(tmp_post_folder, wordpress_url)
file.remove(dir(tmp_post_folder, full.names = TRUE))
}
}
