% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wp_tags}
\alias{wp_tags}
\title{Get WordPress Tags}
\usage{
wp_tags(wordpress_url)
}
\arguments{
\item{wordpress_url}{URL to your website}
}
\value{
A data frame with tag ids and names
}
\description{
This function helps you get your WordPress tags
details so that you can use for your YAML header.
}
\examples{
\dontrun{
wordpress_url <- "https://rmd-wordpress.eu" # replace with your own (test) website
wp_tags(wordpress_url)
}
}
