% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{wp_post}
\alias{wp_post}
\title{Post a post to Wordpress}
\usage{
wp_post(post_folder, wordpress_url)
}
\arguments{
\item{post_folder}{Path to folder where the post (index.Rmd, index.md, media)
lives}

\item{wordpress_url}{URL to your website}
}
\value{
URL to the post (invisibly)
}
\description{
Post a post to Wordpress
}
\examples{
\dontrun{
# this requires authentication!
post_folder <- system.file(file.path("post-example", "index.md"), package = "goodpress")
wordpress_url <- "https://rmd-wordpress.eu" # replace with your own (test) website
wp_post(post_folder, wordpress_url)
}
}
