% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{wp_post}
\alias{wp_post}
\title{Post a post}
\usage{
wp_post(post, wordpress_url)
}
\arguments{
\item{post}{(JSON)}

\item{wordpress_url}{}
}
\value{
URL to the post
}
\description{
Post a post
}
\examples{
  post <- jsonlite::toJSON(
  list( 'date' = '2017-06-19T20:00:35',
        'title' = 'First REST API post',
        'slug' = 'rest-api-1',
        'status' = 'publish',
        'content' = 'this is the <a href="https://masalmon.eu">content</a>',
        'excerpt' = 'Exceptional post!',
        'format' = 'standard'),
  auto_unbox = TRUE
  )
 wordpress_url <- "https://rmd-wordpress.eu"
  wp_post(post, wordpress_url)
}
