% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{wp_post}
\alias{wp_post}
\title{Post a post to Wordpress}
\usage{
wp_post(post_folder, wordpress_url)
}
\arguments{
\item{post_folder}{Path to folder where the post
(index.Rmd with \code{hugodown::md_document} output format,
index.md, images)
lives}

\item{wordpress_url}{URL to your website}
}
\value{
URL to the post (invisibly)
}
\description{
Post a post to Wordpress
}
\details{
Several aspects of the post can be controlled by the YAML of index.Rmd.\if{html}{\out{<div class="yaml">}}\preformatted{---
title: "Title of the Post" # compulsory!
date: "2020-04-01T00:00:00" # compulsory!
author: "admin7891" # Either omit the field (the author will be
#  the authenticated user) or use an username
slug: "post-slug" # Important especially if the slug is in the URL
excerpt: "Here I summarize this fantastic post"
status: "publish" # One of: "publish", "future", "draft",
#  "pending", "private"; "publish" by default
output: hugodown::md_document # Don't change this
categories: # Categories, optional, will be created if
# the post is the first with the category
  - math
- Code and Stuff
tags: # Tags, optional, will be created if
# the post is the first with the tag
  - crul
  - mathjax
  - R packages
comment_status: closed # Either closed (default) or open
ping_status: closed # Either closed (default) or open
---
}\if{html}{\out{</div>}}

The package cannot handle private posts with password, only private posts that
are visible to admins and editors only. You could create a private post, and
then from the WordPress interface make it visible with password. Make it private again
before trying to update the post with the R package.
}
\examples{
\dontrun{
# this requires authentication!
post_folder <- system.file(file.path("post-example"), package = "goodpress")
tmp_post_folder <- file.path(tempdir(), "post")
dir.create(tmp_post_folder)
file.copy(dir(post_folder, full.names = TRUE), tmp_post_folder)
wordpress_url <- "https://rmd-wordpress.eu" # replace with your own (test) website
wp_post(tmp_post_folder, wordpress_url)
file.remove(dir(tmp_post_folder, full.names = TRUE))
}
}
